requirements.txt
SpeechRecognition==3.10.4
pyttsx3==2.90
openai-whisper
transformers
soundfile==0.12.1
replicate
